Add files via upload
Went through some throw statements.  Also applied the bitwise operator " | " for the first time to handle two exceptions in the parameter.
Example: catch (ArithmeticException | NoSuchElementException e) . One use of the try and catch statements is that it allows the user to make mistakes and try again without having to stop the program. After an exception, you can just spit out a message to try again using a while loop. I also used a while(true) loop for the first time, which creates an infinite loop. If I'm not mistaken, isn't that bad practice??